+++ besu/src/test/java/org/hyperledger/besu/cli/BesuCommandTest.java
@@ -3547,29 +3547,12 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void snapsyncHealingOptionShouldBeDisabledByDefault() {
-    final TestBesuCommand besuCommand = parseCommand();
-    assertThat(besuCommand.unstableSynchronizerOptions.isSnapsyncFlatDbHealingEnabled()).isFalse();
-  }
-
-  @Test
-  public void snapsyncHealingOptionShouldWork() {
-    final TestBesuCommand besuCommand =
-        parseCommand("--Xsnapsync-synchronizer-flat-db-healing-enabled", "true");
-    assertThat(besuCommand.unstableSynchronizerOptions.isSnapsyncFlatDbHealingEnabled()).isTrue();
-  }
-
-  @Test
-  public void snapsyncForHealingFeaturesShouldFailWhenHealingIsNotEnabled() {
+  public void snapsyncForHealingFeaturesShouldFailWhenHealingIsNotSet_EnabledByDefault() {
     parseCommand("--Xsnapsync-synchronizer-flat-account-healed-count-per-request", "100");
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains(
-            "--Xsnapsync-synchronizer-flat option can only be used when -Xsnapsync-synchronizer-flat-db-healing-enabled is true");
+    assertThat(commandOutput.toString(UTF_8)).isEmpty();
 
     parseCommand("--Xsnapsync-synchronizer-flat-slot-healed-count-per-request", "100");
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains(
-            "--Xsnapsync-synchronizer-flat option can only be used when -Xsnapsync-synchronizer-flat-db-healing-enabled is true");
+    assertThat(commandOutput.toString(UTF_8)).isEmpty();
   }
 
   @Test